Update QuickSort.cpp
int to double

+++ QuickSort.cpp
@@ -2,25 +2,25 @@
 
 // Model used for quickSort: https://www.geeksforgeeks.org/quick-sort/
 
-void QuickSort::swap(int* a, int* b)
+void QuickSort::swap(double* a, double* b)
 {
-	int t = *a;
+	double t = *a;
 	*a = *b;
 	*b = t;
 }
 
-void QuickSort::quickSort(int array[], int low, int high)
+void QuickSort::quickSort(double array[], int low, int high)
 {
 	if (low < high) {
-		int partI = partition(array, low, high);
+		double partI = partition(array, low, high);
 		quickSort(array, low, partI - 1);
 		quickSort(array, partI + 1, high);
 	}
 }
 
-int QuickSort::partition(int array[], int low, int high)
+int QuickSort::partition(double array[], int low, int high)
 {
-	int pivot = array[high];
+	double pivot = array[high];
 	int i = low - 1;
 	for (int j = low; j <= high; j++) {
 		if (array[j] < pivot) {
